requirements.txt
streamlit==1.35.0
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4
openpyxl==3.1.2
yfinance==0.2.32
python-dateutil==2.8.2
